+++ CHANGES.rst
@@ -6,6 +6,7 @@ Unreleased
 - Add support for `using` chained manager method and save/delete keyword argument (gh-507)
 - Added management command `clean_duplicate_history` to remove duplicate history entries (gh-483)
 - Updated most_recent to work with excluded_fields (gh-477)
+- Fixed bug that prevented self-referential foreign key from using `'self'` (gh-513)
 
 2.6.0 (2018-12-12)
 ------------------